pyproject.toml
[project]
name = "langgraph-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.29",
    "langchain-tavily>=0.2.11",
    "langgraph-cli[inmem]>=0.3.6",
    "openai>=1.99.6",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "langchain-chroma>=0.1.2",
    "pypdf>=4.0.1",
    "python-docx>=1.1.0",
    "jieba==0.42.1",
    "accelerate>=1.10.1",
    "pymupdf>=1.26.4",
    "langchain-huggingface>=0.1.0",
    "langchain-chroma>=0.1.0",
]
